Update About and imports for App.tsx; Update layout for desktop and mobile for both About content and MessageWindow
