+++ frontend/src/Chat.tsx
@@ -16,7 +16,6 @@ export default function Chat() {
       setSelectedStatute(statute);
       setDrawerOpen(true);
     }
-
   };
 
   const closeDrawer = () => {
@@ -29,10 +28,10 @@ export default function Chat() {
       <div className="flex w-full items-center ">
         <div className="flex-1 transition-all duration-300">
           <div
-            className={`container relative flex flex-col mx-auto p-6 bg-[#F4F4F2] rounded-lg shadow-[0_4px_6px_rgba(0,0,0,0.1)]
+            className={`container relative h-[calc(100vh-4rem)] flex flex-col mx-auto p-6 bg-[#F4F4F2] rounded-lg shadow-[0_4px_6px_rgba(0,0,0,0.1)]
               ${
                 isOngoing
-                  ? "justify-between h-[calc(100dvh-10rem)]"
+                  ? "justify-between sm:h-[calc(100dvh-10rem)]"
                   : "justify-center max-w-[600px]"
               }`}
           >
